# Filter excessive attrs from ORCA's DML delete node