… src/backend/gporca/data/dxl/minidump/DeleteWithAfterTriggerDroppedCol.mdp
… src/backend/gporca/data/dxl/minidump/PartitionedDelete.mdp
… src/backend/gporca/libgpopt/src/operators/CPhysicalDML.cpp
… src/backend/gporca/server/src/unittest/gpopt/minidump/CDMLTest.cpp
… src/test/regress/expected/partition_pruning_optimizer.out
… src/test/regress/expected/qp_dropped_cols_optimizer.out