@BenderArenadata 2024-02-20
Allure report https://allure-ee.adsw.io/launch/65027