@BenderArenadata 2024-02-27
Allure report https://allure-ee.adsw.io/launch/65429