@BenderArenadata 2024-02-28
Allure report https://allure-ee.adsw.io/launch/65581