@InnerLife0 2024-03-06
> "source array" is not obvious expression.

Trying to imagine, but can't see a case when it can be not obvious. The comment is given in context of changed code. The code changes are small and source attrs are used only ones. However, I can change a comment if it's a real problem.

> Too informal

Can't find this too informal. But anyway, if yes, is it a real problem to write here like so?

> Why do we really talk about it.

To highlight that we don't ignore this problem. Probably, this should be a part of PR message only, not a commit message. Commit message may have a link to existed PR, if any.